+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Changelog
 - Bump the minimum supported ``hypothesis_graphql`` version to ``0.11.0``.
 - Use the same random seed for all tests in CLI. :issue:`1384`.
 - Improve serialization error messages in CLI.
+- Store skip reason in the runner events.
 
 **Removed**